fix: resolve CI build error and warnings
- SceneRenderer: GetRendering() → Rendering() (correct EngineSettings API)
- SoundEffect: fix DWORD format specifiers (%u → %lu with cast)
- PipelineStateCache: fix uint64_t format specifiers with portable casts
- PrefabManager: fix uint64_t format specifiers with portable casts
- EngineDiagnosticsExtended: fix always-true unsigned >= 0 comparison
- PlatformerCollectibleSystem/PlayerController: add default case for Count enum

+++ SparkEditor/Source/Prefabs/PrefabManager.cpp
@@ -63,7 +63,7 @@ namespace SparkEditor
             else
             {
                 SPARK_LOG_WARN(Spark::LogCategory::Editor, "Entity %llu not found in scene for prefab creation",
-                               entityId);
+                               static_cast<unsigned long long>(entityId));
             }
         }
 
@@ -152,7 +152,7 @@ namespace SparkEditor
         static uint64_t nextFallbackId = 1000;
         uint64_t entityId = nextFallbackId++;
         SPARK_LOG_INFO(Spark::LogCategory::Editor, "Instantiated prefab '%s' as entity %llu (no scene)",
-                       prefabName.c_str(), entityId);
+                       prefabName.c_str(), static_cast<unsigned long long>(entityId));
         RegisterInstance(entityId, prefabName);
         return entityId;
     }
